% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispensingData.R
\docType{data}
\name{dispensingData}
\alias{dispensingData}
\title{Dispensing data.}
\format{
A data frame with 3085 rows and 30 variables:
\describe{
  \item{fips}{FIPS county code}
  \item{State}{State}
  \item{County}{County}
  \item{2006-2020}{estimated opioid dispensing rate per 100 persons in each year}
  \item{POPESTIMATE2010-POPESTIMATE2020}{annual population estimates
  from the US Census Bureau in each year}


  ...
}
}
\source{
\url{https://www.cdc.gov/drugoverdose/rxrate-maps/index.html}
}
\usage{
dispensingData
}
\description{
A dataset containing estimated opioid dispensing rate
per 100 persons in United States, 2006-2020.
}
\keyword{datasets}
