\name{opioidData}
\alias{opioidData}
\docType{data}
\title{Estimated Opioid Dispensing Rates in the USA}
\description{
A dataset containing estimated opioid dispensing rate
per 100 persons in United States, 2006-2020.
}
\usage{opioidData}
\format{
  A data frame with 3085 rows and 18 variables:
  \describe{
    \item{fips}{FIPS county code}
    \item{State}{State}
    \item{County}{County}
    \item{X2006-X2020}{estimated opioid dispensing rate per 100 persons in each year}
  }
}

\source{
\url{https://www.cdc.gov/drugoverdose/rxrate-maps/index.html}
}
\keyword{datasets}
