% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispensingData.R
\docType{data}
\name{opioidData}
\alias{opioidData}
\title{Estimated Opioid Dispensing Rates in the USA}
\format{
A data frame with 3085 rows and 18 variables:
\describe{
  \item{fips}{FIPS county code}
  \item{State}{State}
  \item{County}{County}
  \item{X2006-X2020}{estimated opioid dispensing rate per 100 persons in each year}
}
}
\source{
\url{https://www.cdc.gov/drugoverdose/rxrate-maps/index.html}
}
\usage{
opioidData
}
\description{
A dataset containing estimated opioid dispensing rate
per 100 persons in United States, 2006-2020.
}
\keyword{datasets}
