% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opioidData.R
\docType{data}
\name{opioidData}
\alias{opioidData}
\title{Opioid Dispensing Rates}
\format{
data.frame; columns: fips = FIPS county code, State = State, County = County,
X2006-X2020 = estimated opioid dispensing rate per 100 persons in each year
}
\source{
\url{https://www.cdc.gov/drugoverdose/rxrate-maps/index.html}
}
\usage{
data(opioidData)
}
\description{
A dataset containing estimated opioid dispensing rate per 100 persons in United States, 2006-2020.
}
\keyword{datasets}
