% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsim.R
\name{plotsim}
\alias{plotsim}
\title{plotsim}
\usage{
plotsim(curve)
}
\arguments{
\item{curve}{simulated dataset to be used}
}
\description{
This function generates line plots
based on types of dataset obtained from simcurve().
}
\details{
For dataset containing radomly generated areas with change,
line plots for 1. All areas over time, 2. areas with change, and
3. areas with no change are generated.
For dataset containing curves with fixed coefficients and random noise,
line plots of simulated data for 3 clusters are generated.
}
